ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
pytesseract
Pillow
